refactor(server): extract Zeus.Server.Hosting library
Move the entire host pipeline — services, route map, Kestrel/DI/middleware
config — out of Zeus.Server/Program.cs (~1200 LoC of top-level statements)
into a new Microsoft.NET.Sdk library project. Public API on the library is
ZeusHost.Build / InitializeAsync / RunAsync plus a ZeusHostOptions surface
covering port, bind policy, HTTPS, and console banner.

Zeus.Server/Program.cs is now a ~25-line thunk that builds service-mode
options (port from ZEUS_PORT or 6060, bind all interfaces, HTTPS via the
LAN cert, banner on) and calls ZeusHost.RunAsync. The 'public partial
class Program;' anchor is preserved so Microsoft.AspNetCore.Mvc.Testing's
WebApplicationFactory<Program> in Zeus.Server.Tests still hooks the
endpoint pipeline.

All 36 service files keep their original Zeus.Server / Zeus.Server.Tci
namespaces, so existing 'using Zeus.Server;' in tests stays green.
appsettings.*.json, wdsp-data/zetaHat.bin, wdsp-data/calculus, and the
SPA wwwroot now ship from the library via Content/CopyToPublishDirectory
so any consumer (Zeus.Server today, Zeus.Desktop next) picks them up
without duplication.

Vite outDir + .gitignore updated to track the new wwwroot location.

dotnet build Zeus.slnx: 0 errors, 0 warnings.
dotnet test Zeus.slnx:  575 passed, 0 failed (incl. 226 Zeus.Server.Tests).
