+++ Zeus.Server.Hosting/CredentialStore.cs
