+++ Zeus.Server.Hosting/LogService.cs
