+++ Zeus.Server.Hosting/Tci/SpotManager.cs
