+++ Zeus.Server.Hosting/Zeus.Server.Hosting.csproj
@@ -1,0 +1,70 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <SpaRoot>$(MSBuildThisFileDirectory)..\zeus-web</SpaRoot>
+    <SpaBuildOutput>$(MSBuildThisFileDirectory)wwwroot\index.html</SpaBuildOutput>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
+
+  <!-- Mirror the implicit usings that Microsoft.NET.Sdk.Web injects automatically.
+       Required because we use the plain Sdk (so the assembly can be referenced
+       by non-web consumers like Zeus.Desktop without dragging in the Web SDK
+       constraints) but still rely on ASP.NET Core types pervasively. -->
+  <ItemGroup>
+    <Using Include="System.Net.Http.Json" />
+    <Using Include="Microsoft.AspNetCore.Builder" />
+    <Using Include="Microsoft.AspNetCore.Hosting" />
+    <Using Include="Microsoft.AspNetCore.Http" />
+    <Using Include="Microsoft.AspNetCore.Routing" />
+    <Using Include="Microsoft.Extensions.Configuration" />
+    <Using Include="Microsoft.Extensions.DependencyInjection" />
+    <Using Include="Microsoft.Extensions.Hosting" />
+    <Using Include="Microsoft.Extensions.Logging" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Zeus.Contracts\Zeus.Contracts.csproj" />
+    <ProjectReference Include="..\Zeus.Protocol1\Zeus.Protocol1.csproj" />
+    <ProjectReference Include="..\Zeus.Protocol2\Zeus.Protocol2.csproj" />
+    <ProjectReference Include="..\Zeus.Dsp\Zeus.Dsp.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="Zeus.Contracts.Tests" />
+    <InternalsVisibleTo Include="Zeus.Server.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="LiteDB" Version="5.0.21" />
+  </ItemGroup>
+
+  <!-- WDSP NR2 EMNR fopen()s these by bare relative name (native/wdsp/emnr.c:215,397).
+       Link= drops them at the output-dir root of any consumer (Zeus.Server, Zeus.Desktop)
+       so the cwd-anchor in ZeusHost.cs finds them. -->
+  <ItemGroup>
+    <Content Include="wdsp-data\zetaHat.bin" Link="zetaHat.bin" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" />
+    <Content Include="wdsp-data\calculus" Link="calculus" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" />
+  </ItemGroup>
+
+  <!-- appsettings flow to consumers — preserves the existing 'next to the
+       executable' lookup that WebApplication.CreateBuilder does by default. -->
+  <ItemGroup>
+    <Content Include="appsettings.json" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" />
+    <Content Include="appsettings.Development.json" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" />
+  </ItemGroup>
+
+  <!-- SPA assets are produced by Zeus.Server's BuildSpa target on publish; for
+       library consumers (Zeus.Desktop) we recursively copy whatever's in
+       wwwroot/ at consume time. The actual build is owned by the executable
+       project so 'dotnet build' on a library never shells out to npm. -->
+  <ItemGroup>
+    <Content Include="wwwroot\**\*" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" Condition="Exists('$(MSBuildThisFileDirectory)wwwroot')" />
+  </ItemGroup>
+
+</Project>
