+++ Zeus.Server/Zeus.Server.csproj
@@ -1,38 +1,23 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <ItemGroup>
-    <ProjectReference Include="..\Zeus.Contracts\Zeus.Contracts.csproj" />
-    <ProjectReference Include="..\Zeus.Protocol1\Zeus.Protocol1.csproj" />
-    <ProjectReference Include="..\Zeus.Protocol2\Zeus.Protocol2.csproj" />
-    <ProjectReference Include="..\Zeus.Dsp\Zeus.Dsp.csproj" />
-  </ItemGroup>
-
   <PropertyGroup>
     <TargetFramework>net10.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <SpaRoot>$(MSBuildThisFileDirectory)..\zeus-web</SpaRoot>
-    <SpaBuildOutput>$(MSBuildThisFileDirectory)wwwroot\index.html</SpaBuildOutput>
+    <SpaBuildOutput>$(MSBuildThisFileDirectory)..\Zeus.Server.Hosting\wwwroot\index.html</SpaBuildOutput>
   </PropertyGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Zeus.Contracts.Tests" />
-    <InternalsVisibleTo Include="Zeus.Server.Tests" />
+    <ProjectReference Include="..\Zeus.Server.Hosting\Zeus.Server.Hosting.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="LiteDB" Version="5.0.21" />
-  </ItemGroup>
-
-  <!-- WDSP NR2 EMNR fopen()s these by bare relative name (native/wdsp/emnr.c:215,397).
-       Link= drops them at the output-dir root so the cwd-anchor in Program.cs finds them. -->
-  <ItemGroup>
-    <Content Include="wdsp-data\zetaHat.bin" Link="zetaHat.bin" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" />
-    <Content Include="wdsp-data\calculus" Link="calculus" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="PreserveNewest" />
-  </ItemGroup>
-
+  <!-- Build the React frontend into Zeus.Server.Hosting/wwwroot so it flows to
+       every downstream consumer (Zeus.Server publish output, Zeus.Desktop, etc).
+       Owning the SPA build at the executable layer (not in the library) keeps
+       'dotnet build Zeus.Server.Hosting' from shelling out to npm. -->
   <Target Name="BuildSpa" BeforeTargets="Publish" Condition="!Exists('$(SpaBuildOutput)') Or '$(ForceSpaBuild)' == 'true'">
-    <Message Importance="high" Text="Building React frontend in $(SpaRoot)" />
+    <Message Importance="high" Text="Building React frontend in $(SpaRoot) → $(SpaBuildOutput)" />
     <Exec Command="npm ci" WorkingDirectory="$(SpaRoot)" />
     <Exec Command="npm run build" WorkingDirectory="$(SpaRoot)" />
   </Target>
