+++ Zeus.slnx
@@ -11,4 +11,5 @@
   <Project Path="Zeus.Protocol1/Zeus.Protocol1.csproj" />
   <Project Path="Zeus.Protocol2/Zeus.Protocol2.csproj" />
   <Project Path="Zeus.Server/Zeus.Server.csproj" />
+  <Project Path="Zeus.Server.Hosting/Zeus.Server.Hosting.csproj" />
 </Solution>
